# projeto_Renovacao/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "API de Gerenciamento de Empresas, Cursos e Demandas",
      "description": "API para gerenciamento de empresas, instituições, cursos, professores e demandas, com funcionalidades CRUD e integração de processos educacionais e empresariais.",
      "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
"/demandasPegas/{id}": {
  "put": {
    "tags": ["DemandaPega"],
    "summary": "Atualizar as demandas pegas",
    "description": "Este endpoint atualiza as demandas pegas.",
    "operationId": "AtualizarDemandaPegas",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "description": "ID da demanda pega que será atualizada.",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    ],
    "requestBody": {
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "descricao": {
                "type": "string",
                "description": "Atualização da descrição do professor(a)."
              }
            },
            "required": ["descricao"]
          }
        }
      }
    },
    "responses": {
      "204": {
        "description": "Demanda pega atualizada com sucesso, sem corpo na requisição"
      }
      ,
      "500":{
         "description": "Erro no servidor interno."
      }
    }
  },
  "patch":{
    "tags": ["DemandaPega"],
    "summary": "Entregar demanda",
    "description": "Este endpoint atualiza data de entrega.",
    "operationId": "Entregar",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "description": "ID da demanda pega que será entregue.",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    ],
    "responses":{
"204": {
        "description": "Demanda pega entregue com sucesso, sem corpo na requisição"
      }
      ,
      "500":{
         "description": "Erro no servidor interno."
      }
    }
  },
  "delete":{
    "tags": ["DemandaPega"],
    "summary": "Deletar demanda pega",
    "description": "Este endpoint deleta a demanda pega pelo id.",
    "operationId": "deletarDemandaPega",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "description": "ID da demanda pega que será deletada.",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    ],
    "responses":{
      "204": {
        "description": "Demanda pega entregue com sucesso, sem corpo na requisição"
      }
      ,
      "500":{
         "description": "Erro no servidor interno."
      }
    }
  }
},
"/demandasPegas/professores":{
"get":{
  "tags": ["DemandaPega"],
    "summary": "Listar com demandas pegas e professores(a)",
    "description": "Este endpoint retorna uma lista com demandas que foram pegas e professores(a) a pegaram.",
    "operationId": "mostrarDemandasProfessores",
    "responses":{
      "200":{
        "description": "Todas as demandas pegas com sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type":"object",
              "properties":{
                "professor":{
                  "items": {
                    "type":"array",
                      "$ref": "#/components/schemas/Professor"
                    
                  }
                },
                "demandas":{
                  "items": {
                    "type":"array",
                      "$ref": "#/components/schemas/Demanda"
                    
                  }
                },
                "demandasPega":{
                  "items": {
                    "type":"array",
                      "$ref": "#/components/schemas/DemandaPega"
                    
                  }
                }
                
              }
             
            }
          }
        }
      },
      "500":{
         "description": "Erro no servidor interno."
      }
    }
}
},
"/demandasPegas/demandas/{id}": {
  "get": {
    "tags": ["DemandaPega"],
    "summary": "Mostra todos os professores que pegaram o ID da demanda",
    "description": "Este endpoint retorna um objeto contendo a demanda especificada e uma lista de professores que pegaram essa demanda.",
    "operationId": "mostrarProfessoresPorDemanda",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "description": "ID da demanda para buscar os professores relacionados.",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    ],
    "responses": {
      "200": {
        "description": "A demanda e os professores associados retornados com sucesso.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "demanda": {
                  "$ref": "#/components/schemas/Demanda"
                },
                "professores": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Professor"
                  }
                }
              }
            }
          }
        }
      },
      "500":{
         "description": "Erro no servidor interno."
      }
    }
  }
},
"/demandasPegas/professores/{id}":{

  "get":{
  
      "tags": ["DemandaPega"],
      "summary": "Mostra todas demandas que o professor(a) pegou",
      "description": "Este endpoint retorna um objeto contendo o professor especificada e uma lista de demandas que Id está relacionado.",
      "operationId": "mostrarDemandasPorProfessor",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "ID do professor para buscar as demandas relacionados.",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "A demanda e os professores associados retornados com sucesso.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "professor": {
                    "$ref": "#/components/schemas/Professor"
                  },
                  "demandas": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Demanda"
                    }
                  }
                }
              }
            }
          }
        },
        "500":{
         "description": "Erro no servidor interno."
      }
      }
    
  }
},
...
"/cursos/{id}": {
  "patch": {
    "tags": ["Cursos"],
    "summary": "Atualizar o nome de um curso",
    "description": "Este endpoint permite atualizar o nome de um curso existente.",
    "operationId": "atualizarCurso",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "description": "ID do curso que será atualizado.",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    ],
    "requestBody": {
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "novo_nome": {
                "type": "string",
                "description": "Novo nome do curso."
              }
            },
            "required": ["novo_nome"]
          }
        }
      }
    },
    "responses": {
      "204": {
        "description": "Nome do Curso atualizado com sucesso, sem corpo na resposta"
      },
      "500":{
         "description": "Erro no servidor interno."
      }
    }
  },
  "delete": {
    "tags": ["Cursos"],
    "summary": "Deletar um curso",
    "description": "Este endpoint permite deletar um curso existente.",
    "operationId": "deletarCurso",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "description": "ID do curso que será deletado.",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    ],
    "responses": {
      "204": {
        "description": "Curso deletado com sucesso."
      },
      "500":{
         "description": "Erro no servidor interno."
      }
    }
  }
},
...
  },
  "components": {
    "schemas": {
...
        "Demanda": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "ID da demanda"
                },
                "titulo": {
                    "type": "string",
                    "description": "Título da demanda"
                },
                "descricao": {
                    "type": "string",
                    "description": "Descrição da demanda"
                },
                "dataEnvio": {
                    "type": "string",
                    "format": "date",
                    "description": "Data de envio da demanda"
                },
                "dataLimiteParaFicarDisponivel": {
                    "type": "string",
                    "format": "date",
                    "description": "Data limite para a demanda ficar disponível"
                },
                "prazo": {
                    "type": "string",
                    "description": "Prazo da demanda"
                }
            }
            
        },
...
         "Professor":{
          "type":"object",
          "properties":{
            "id": {
              "type": "integer",
              "description": "ID do professor(a)"
          },
          "nome": {
              "type": "string",
              "description": "Nome do professor(a)"
          },
          "email": {
            "type": "string",
            "description": "Nome do professor(a)"
        }

          }
         },
         "DemandaPega":{
          "type":"object",
          "properties":{
            "id": {
              "type": "integer",
              "description": "ID da demanda pega"
          },
          "status": {
            "type": "string",
            "description": "Status da demanda pega"
        },
        "descricão": {
          "type": "string",
          "description": "Descrição da demanda pega"
      },  
      "data_ultima_atualizacao": {
        "type": "string",
        "description": "Data da ultima da atualizacao da demanda que foi pega"
    },
    "data_demanda_pega": {
        "type": "string",
        "description": "Data que demanda foi pega"
    },
    "data_entrega": {
        "type": "string",
        "description": "Data que demanda foi entregue"
    },"prazo": {
        "type": "string",
        "description": "Data do prazo para entregar a demanda pega"
    }
          }
         }
    }
}
}